[parser] feat: added more granluar "interfaces" for productions
included are: Expressions, Unit, IdentifierProds, and Collection.
Expressions are those with operands; Units are those that evaluate to a
single value but is not an Expression; IdentifierProds are identifiers
that are either used to index into an array, fn call, or access class
properties/methods; lastly, Collections are the arrays and string fmt.